Dev: move cctp deposit functions into starter
